requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
objects==0.3.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
unicodecsv==0.14.1
